requirements.txt
streamlit
hopsworks
scikit-learn
tensorflow
shap
pandas
requests
xgboost
python-dotenv
pyyaml
